+++ prisma/migrations/20260430211717_init/migration.sql
@@ -84,8 +84,6 @@ CREATE TABLE "sessions" (
     "revokedAt" TIMESTAMP(3),
     "revokedBy" TEXT,
     "deviceId" TEXT NOT NULL,
-    "ipAddress" TEXT,
-    "userAgent" TEXT,
     "lastUsedAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updatedAt" TIMESTAMP(3) NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,